+++ Runtime/Pools/QueuePool.cs
@@ -1,31 +1,33 @@
+using System.Collections;
 using System.Collections.Generic;
+using UnityEngine;
 
 namespace Unbegames.Services {
   /// <summary>
-  /// HashSet Pool.
+  /// Queue Pool.
   /// </summary>
   /// <typeparam name="T">Type of the objects in the pooled hashsets.</typeparam>
-  public static class HashSetPool<T> {
+  public static class QueuePool<T> {
     // Object pool to avoid allocations.
-    static readonly ObjectPool<HashSet<T>> pool = new ObjectPool<HashSet<T>>(null, h => h.Clear());
+    static readonly ObjectPool<Queue<T>> pool = new ObjectPool<Queue<T>>(null, q => q.Clear());
 
     /// <summary>
-    /// Get a new HashSet
+    /// Get a new Queue
     /// </summary>
     /// <returns>A new HashSet</returns>
-    public static HashSet<T> Get() => pool.Get();
+    public static Queue<T> Get() => pool.Get();
 
     /// <summary>
-    /// Get a new HashSet PooledObject.
+    /// Get a new queue PooledObject.
     /// </summary>
     /// <param name="value">Output typed HashSet.</param>
     /// <returns>A new HashSet PooledObject.</returns>
-    public static ObjectPool<HashSet<T>>.PooledObject Get(out HashSet<T> value) => pool.Get(out value);
+    public static ObjectPool<Queue<T>>.PooledObject Get(out Queue<T> value) => pool.Get(out value);
 
     /// <summary>
-    /// Release an HashSet to the pool.
+    /// Release the queue to the pool.
     /// </summary>
     /// <param name="toRelease">hashSet to release.</param>
-    public static void Release(HashSet<T> toRelease) => pool.Release(toRelease);
+    public static void Release(Queue<T> toRelease) => pool.Release(toRelease);
   }
 }