+++ CHANGELOG.md
@@ -1,7 +1,15 @@
+## [1.1.1] - 2022-01-14
+
+### Added
+- QueuePool
+
+### Fixed
+- Fixed missing HashSet pool
+
 ## [1.1.0] - 2022-01-14
 
 ### Added
-- Object, List, Dictionary and HashMap pools.
+- Object, List, Dictionary and HashSet pools.
 
 ## [1.0.0] - 2022-01-11